Simulation GUI front end.
xsim.[cpp|h] contains XNernstSim, which inherits from both
QWidget (for all of the qt slots, signals, and graphics) and
NernstSim (in sim.[cpp.h]).  If you need to add something to
the gui, do it here (or in paint.cpp, or gui.pp).  If you
need to add something to the simulation engine, do it in
sim.cpp.